Adds JWT authentication, auditing, Spring Security, OpenAPI, and restructures the project.

+++ leilaSalao/src/main/java/com/leila/leilaSalao/repository/UserRepository.java
@@ -1,9 +1,11 @@
 package com.leila.leilaSalao.repository;
 
-import com.leila.leilaSalao.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.leila.leilaSalao.model.User;
+
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
 }
